+++ src/player/Player.hpp
@@ -4,6 +4,7 @@
 # include <memory>
 # include "Strategy.hpp"
 # include <vector>
+# include <cstdint>
 
 class Player {
     private:
@@ -14,11 +15,11 @@ class Player {
         int                             _turnsLeftInJail;
         unsigned int                    _doublesRolledInARow;
 
-        unsigned int                    _currentMoney;
+        int64_t                         _currentMoney;
         std::vector<unsigned int>       _squaresOwned;
         std::shared_ptr<const Strategy> _strategy;
     public:
-        Player(unsigned int id, const std::string &name, unsigned int startingSquare, unsigned int startingMoney);
+        Player(unsigned int id, const std::string &name, unsigned int startingSquare, int64_t startingMoney);
         void setCurrentSquare(unsigned int square);
         unsigned int getCurrentSquare() const;
